# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tal Misri Factory Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties:
    get:
      operationId: listParties
      tags: [parties]
      summary: List all parties (customers/suppliers)
      responses:
        "200":
          description: List of parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Party"
    post:
      operationId: createParty
      tags: [parties]
      summary: Create a new party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartyBody"
      responses:
        "201":
          description: Created party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{id}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get a party with full ledger entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyDetail"
    put:
      operationId: updateParty
      tags: [parties]
      summary: Update a party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartyBody"
      responses:
        "200":
          description: Updated party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    delete:
      operationId: deleteParty
      tags: [parties]
      summary: Delete a party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ledger:
    get:
      operationId: listLedgerEntries
      tags: [ledger]
      summary: List ledger entries for a party
      parameters:
        - name: partyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
    post:
      operationId: createLedgerEntry
      tags: [ledger]
      summary: Create a new ledger entry for a party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerEntryBody"
      responses:
        "201":
          description: Created ledger entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
  /ledger/{id}:
    delete:
      operationId: deleteLedgerEntry
      tags: [ledger]
      summary: Delete a ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── PRODUCTION BATCHES ──
...
  /stock/transactions:
    get:
      operationId: listStockTransactions
      tags: [stock]
      summary: List stock transactions
      parameters:
        - name: itemType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stock transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockTransaction"
    post:
      operationId: createStockTransaction
      tags: [stock]
      summary: Add a stock transaction (in/out)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockTransaction"

  # ── BALANCE SHEET ──
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Party:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameBn:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        type:
          type: string
          enum: [customer, supplier, both, loan_given, loan_taken, other]
        balance:
          type: number
        lastTransactionDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, type, balance, createdAt]
    PartyDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameBn:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        type:
          type: string
          enum: [customer, supplier, both, loan_given, loan_taken, other]
        balance:
          type: number
        lastTransactionDate:
          type: string
          nullable: true
        createdAt:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
      required: [id, name, type, balance, createdAt, entries]
    CreatePartyBody:
      type: object
      properties:
        name:
          type: string
        nameBn:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        type:
          type: string
          enum: [customer, supplier, both, loan_given, loan_taken, other]
      required: [name, type]
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        partyId:
          type: integer
        date:
          type: string
        type:
          type: string
          enum: [debit, credit]
        amount:
          type: number
        description:
          type: string
        itemType:
          type: string
          nullable: true
          enum: [grade1, grade2, byproduct, null]
        itemQtyKg:
          type: number
          nullable: true
        pricePerKg:
          type: number
          nullable: true
        runningBalance:
          type: number
        createdAt:
          type: string
      required: [id, partyId, date, type, amount, description, runningBalance, createdAt]
    CreateLedgerEntryBody:
      type: object
      properties:
        partyId:
          type: integer
        date:
          type: string
        type:
          type: string
          enum: [debit, credit]
        amount:
          type: number
        description:
          type: string
        itemType:
          type: string
          nullable: true
          enum: [grade1, grade2, byproduct, null]
        itemQtyKg:
          type: number
          nullable: true
        pricePerKg:
          type: number
          nullable: true
      required: [partyId, date, type, amount, description]
...
    StockTransaction:
      type: object
      properties:
        id:
          type: integer
        itemType:
          type: string
          enum: [sugar, grade1, grade2, byproduct, sala, gas]
        txType:
          type: string
          enum: [in, out]
        quantityKg:
          type: number
        date:
          type: string
        note:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, itemType, txType, quantityKg, date, createdAt]
    CreateStockTransactionBody:
      type: object
      properties:
        itemType:
          type: string
          enum: [sugar, grade1, grade2, byproduct, sala, gas]
        txType:
          type: string
          enum: [in, out]
        quantityKg:
          type: number
        date:
          type: string
        note:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
      required: [itemType, txType, quantityKg, date]
...