# lib/api-zod/src/generated/types/createLedgerEntryBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Tal Misri Factory Business Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLedgerEntryBodyItemType } from './createLedgerEntryBodyItemType';
import type { CreateLedgerEntryBodyType } from './createLedgerEntryBodyType';

export interface CreateLedgerEntryBody {
  partyId: number;
  date: string;
  type: CreateLedgerEntryBodyType;
  amount: number;
  description: string;
  itemType?: CreateLedgerEntryBodyItemType;
  itemQtyKg?: number | null;
  pricePerKg?: number | null;
}