# lib/api-zod/src/generated/types/createPartyBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Tal Misri Factory Business Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePartyBodyType } from './createPartyBodyType';

export interface CreatePartyBody {
  name: string;
  nameBn?: string | null;
  phone?: string | null;
  address?: string | null;
  type: CreatePartyBodyType;
}