# lib/api-zod/src/generated/types/party.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Tal Misri Factory Business Management API
 * OpenAPI spec version: 0.1.0
 */
import type { PartyType } from './partyType';

export interface Party {
  id: number;
  name: string;
  nameBn?: string | null;
  phone?: string | null;
  address?: string | null;
  type: PartyType;
  balance: number;
  lastTransactionDate?: string | null;
  createdAt: string;
}